# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "My API",
        "description": "REST API for the sonotrak application."
    },
...
    "host": "localhost:8080",
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "definitions": {
...
        "calendar": {
            "type": "object",
            "properties": {
                "monday": {
                    "type": "integer"
                },
                "tuesday": {
                    "type": "integer"
                },
                "wednesday": {
                    "type": "integer"
                },
                "thursday": {
                    "type": "integer"
                },
                "friday": {
                    "type": "integer"
                },
                "saturday": {
                    "type": "integer"
                },
                "sunday": {
                    "type": "integer"
                },
                "start_date": {
                    "type": "string"
                },
                "end_date": {
                    "type": "string"
                }
            }
        },
        "route": {
            "type": "object",
            "properties": {
                "agency_id": {
                    "type": "string"
                },
                "route_short_name": {
                    "type": "string"
                },
                "route_long_name": {
                    "type": "string"
                },
                "route_desc": {
                    "type": "string"
                },
                "route_type": {
                    "type": "integer"
                },
                "route_url": {
                    "type": "string"
                },
                "route_color": {
                    "type": "string"
                },
                "route_text_color": {
                    "type": "string"
                }
            }
        },
...
    },
    "paths": {
...
        "/api/calendar/{id}": {
            "get": {
                "summary": "get calendar by id ",
                "description": "Updates an existing calendar with new information",
                "tags": [
                    "Calendar"
                ],
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/calendar"
                        }
                    },
                    "401": {
                        "description": "Unauthenticated!"
                    }
                }
            },
            "put": {
                "summary": "Update an existing calendar",
                "description": "Updates an existing calendar with new information",
                "tags": [
                    "Calendar"
                ],
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "user",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/calendar"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/calendar"
                        }
                    },
                    "401": {
                        "description": "Unauthenticated!"
                    }
                }
            },
            "delete": {
                "summary": "delete calendar by id ",
                "description": "delete an existing calendar",
                "tags": [
                    "Calendar"
                ],
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/calendar"
                        }
                    },
                    "401": {
                        "description": "Unauthenticated!"
                    }
                }
            }
        },
...
        "/api/route/{id}": {
            "get": {
                "summary": "get route by id ",
                "description": "Updates an existing route with new information",
                "tags": [
                    "Route"
                ],
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/route"
                        }
                    },
                    "401": {
                        "description": "Unauthenticated!"
                    }
                }
            },
            "put": {
                "summary": "Update an existing route",
                "description": "Updates an existing route with new information",
                "tags": [
                    "Route"
                ],
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "user",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/route"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/route"
                        }
                    },
                    "401": {
                        "description": "Unauthenticated!"
                    }
                }
            },
            "delete": {
                "summary": "delete route by id ",
                "description": "delete an existing route",
                "tags": [
                    "Route"
                ],
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/route"
                        }
                    },
                    "401": {
                        "description": "Unauthenticated!"
                    }
                }
            }
        },
...
    }
}